Migrate expect(Deprecation|Notice|Warning) to the PHPUnit8 compat (where it belongs) and add the rest of the related methods.
